pyproject.toml

[project]
name = "audio-compositor-adk"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "python-dotenv",
    "mcp",
    "litellm",
]
